+++ unit-tests/libtest/tools/can-send.hpp
@@ -38,11 +38,11 @@ class CanSend {
 
   static int ret;
   static void* user_data;
-  static int bus_id;
+  static uint_least8_t bus_id;
   static can_msg msg;
   static can_msg* msg_buf;
 
-  static int Func(const can_msg* msg_, int bus_id_, void* data_);
+  static int Func(const can_msg* msg_, uint_least8_t bus_id_, void* data_);
   static void CheckMsg(uint_least32_t id, uint_least8_t flags,
                        uint_least8_t len, const uint_least8_t* data);
   static void Clear();
